+++ feature/product/delivery/response.go
@@ -3,12 +3,12 @@ package delivery
 import "commerce-app/domain"
 
 type ProductResponse struct {
-	ID          int     `json:"id"`
-	Name        string  `json:"name" form:"name"`
-	Price       float64 `json:"price" form:"price"`
-	Stock       int32   `json:"stock" form:"stock"`
-	Description string  `json:"desription" form:"desription"`
-	Images      string  `json:"images_url" form:"images_url"`
+	ID          int    `json:"id"`
+	Name        string `json:"name" form:"name"`
+	Price       int32  `json:"price" form:"price"`
+	Stock       int32  `json:"stock" form:"stock"`
+	Description string `json:"desription" form:"desription"`
+	Images      string `json:"images_url" form:"images_url"`
 }
 
 func FromDomain(data domain.Product) ProductResponse {
@@ -16,7 +16,7 @@ func FromDomain(data domain.Product) ProductResponse {
 	res.ID = int(data.ID)
 	res.Name = data.Name
 	res.Images = data.Images
-	res.Price = data.Price
+	res.Price = int32(data.Price)
 	res.Stock = int32(data.Stock)
 	return res
 }
